generatore_preventivi_py/requirements.txt
customtkinter
python-dotenv
google-generativeai
pypdf